feat(US-9.4): Seed inventory tab view with batch operations
- Add SeedInventoryView as third main-window tab (Ctrl+3)
- Summary stats bar: total packets, expired count, out-of-stock count
- Multi-selection table with search, status and year filters
- Quick-add button; Edit (single selection); Mark as Used + Delete Selected batch ops
- Remove redundant Plants → Manage Seed Inventory menu item (tab replaces it)
- Rename _SeedTableModel → SeedTableModel (public, shared with view)
- Remove now-orphaned SeedInventoryDialog class, its exports, tests and translations
- Update EN + DE translations; recompile .qm files
